revert back to original processing
